+++ isitfit/mainManager.py
@@ -19,6 +19,7 @@ SECONDS_IN_ONE_DAY = 60*60*24 # 86400  # used for granularity (daily)
 MINUTES_IN_ONE_DAY = 60*24 # 1440
 N_DAYS=90
 
+from .utils import IsitfitError
 
 class NoCloudtrailException(Exception):
     pass
@@ -74,7 +75,7 @@ class MainManager:
 
     def add_listener(self, event, listener):
       if event not in self.listeners:
-        raise ValueError("Event %s is not supported for listeners. Use: %s"%(event, ",".join(self.listeners.keys())))
+        raise IsitfitError("Internal dev error: Event %s is not supported for listeners. Use: %s"%(event, ",".join(self.listeners.keys())))
 
       self.listeners[event].append(listener)
 
@@ -215,7 +216,7 @@ class MainManager:
           return pd.DataFrame() # use an empty dataframe in order to distinguish when getting from cache if not available in cache or data not found but set in cache
 
         if len(df_cw1) >1:
-          raise ValueError("More than 1 cloudwatch metric found for %s"%ec2_obj.instance_id)
+          raise IsitfitError("More than 1 cloudwatch metric found for %s"%ec2_obj.instance_id)
 
         # merge
         # df_cw2 = pd.concat(df_cw1, axis=1)