+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Semantic versioning
 Version latest (0.5.0?, 2019-09-24?)
 
 - ...
+- enh: replace ValueError with IsitfitError
 
 
 Version 0.8.4 (2019-10-04)